This is not a converter so let's move this to a different file
