+++ redbot/cogs/cleanup/converters.py
@@ -1,7 +1,6 @@
 from typing import NewType, TYPE_CHECKING
 
-from redbot.core.commands import BadArgument, Context, Converter, permissions_check
-from redbot.core.utils.mod import is_mod_or_superior, check_permissions
+from redbot.core.commands import BadArgument, Context, Converter
 from redbot.core.i18n import Translator
 from redbot.core.utils.chat_formatting import inline
 
@@ -29,16 +28,3 @@ else:
         if ret <= 0:
             raise BadArgument(_("{arg} is not a positive integer.").format(arg=inline(arg)))
         return ret
-
-
-def check_self_permissions():
-    async def predicate(ctx: Context):
-        if not ctx.guild:
-            return True
-        if await check_permissions(ctx, {"manage_messages": True}) or await is_mod_or_superior(
-            ctx.bot, ctx.author
-        ):
-            return True
-        return False
-
-    return permissions_check(predicate)
